Ajustando o tipo do Valor de Ingresso 2

+++ database/migrations/2022_06_23_230231_create_unidades_table.php
@@ -17,7 +17,7 @@ class CreateUnidadesTable extends Migration
             $table->id();
             $table->string('nome',60);
             $table->integer('quantidade');
-            $table->integer('valorIngresso');
+            $table->double('valorIngresso');
             $table->string('descricao',250);
             $table->string('endereco',250);
             $table->timestamps();
